Cleaned up unit test files
