--- test/test_._example.c
@@ -1,6 +1,0 @@
-#include "unity.h"
-#include "._example.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
