--- test/test_._port_rp2350.c
@@ -1,6 +1,0 @@
-#include "unity.h"
-#include "._port_rp2350.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
