--- test/test_._port_stm32.c
@@ -1,6 +1,0 @@
-#include "unity.h"
-#include "._port_stm32.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
