--- test/test_._port_unix.c
@@ -1,6 +1,0 @@
-#include "unity.h"
-#include "._port_unix.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
