--- test/test_._port_windows.c
@@ -1,6 +1,0 @@
-#include "unity.h"
-#include "._port_windows.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
