--- test/test_CMakeCCompilerId.c
@@ -1,11 +1,0 @@
-#include "unity.h"
-#include "CMakeCCompilerId.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_main_should_DoSomething(void) {
-    main();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
