--- test/test_example.c
@@ -1,21 +1,0 @@
-#include "unity.h"
-#include "example.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_OnReceiveCallback_should_DoSomething(void) {
-    OnReceiveCallback();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_OnSendCallback_should_DoSomething(void) {
-    OnSendCallback();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_main_should_DoSomething(void) {
-    main();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
