--- test/test_mbedtls_wrapper.c
@@ -1,56 +1,0 @@
-#include "unity.h"
-#include "mbedtls_wrapper.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_dumpBuf_should_DoSomething(void) {
-    dumpBuf();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_dumpPubkey_should_DoSomething(void) {
-    int result = dumpPubkey();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_mbSha256Buffer_should_DoSomething(void) {
-    mbSha256Buffer();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_mbAes256Buffer_should_DoSomething(void) {
-    mbAes256Buffer();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_rawToDer_should_DoSomething(void) {
-    rawToDer();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_derToRaw_should_DoSomething(void) {
-    derToRaw();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_if_should_DoSomething(void) {
-    if();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_if_should_DoSomething(void) {
-    if();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_mbSignSha256_should_DoSomething(void) {
-    int result = mbSignSha256();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_mbVerifySignatureSecp256K1_should_DoSomething(void) {
-    mbVerifySignatureSecp256K1();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
