--- test/test_picoboot_connection.c
@@ -1,151 +1,0 @@
-#include "unity.h"
-#include "picoboot_connection.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_crc32ForByte_should_DoSomething(void) {
-    crc32ForByte();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_crc32Sw_should_DoSomething(void) {
-    crc32Sw();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_picobootOpenDevice_should_DoSomething(void) {
-    picobootOpenDevice();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_if_should_DoSomething(void) {
-    int result = if();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_if_should_DoSomething(void) {
-    int result = if();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_isHalted_should_DoSomething(void) {
-    int result = isHalted();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootReset_should_DoSomething(void) {
-    int result = picobootReset();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootCmdStatusVerbose_should_DoSomething(void) {
-    int result = picobootCmdStatusVerbose();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootCmdStatus_should_DoSomething(void) {
-    int result = picobootCmdStatus();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootCmd_should_DoSomething(void) {
-    int result = picobootCmd();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_switch_should_DoSomething(void) {
-    switch();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_switch_should_DoSomething(void) {
-    switch();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_picobootExclusiveAccess_should_DoSomething(void) {
-    int result = picobootExclusiveAccess();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootExitXip_should_DoSomething(void) {
-    int result = picobootExitXip();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootEnterCmdXip_should_DoSomething(void) {
-    int result = picobootEnterCmdXip();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootReboot_should_DoSomething(void) {
-    int result = picobootReboot();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootReboot2_should_DoSomething(void) {
-    int result = picobootReboot2();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootExec_should_DoSomething(void) {
-    int result = picobootExec();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootExec2_should_DoSomething(void) {
-    int result = picobootExec2();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootFlashErase_should_DoSomething(void) {
-    int result = picobootFlashErase();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootVector_should_DoSomething(void) {
-    int result = picobootVector();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootWrite_should_DoSomething(void) {
-    int result = picobootWrite();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootRead_should_DoSomething(void) {
-    picobootRead();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_picobootOtpWrite_should_DoSomething(void) {
-    int result = picobootOtpWrite();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootOtpRead_should_DoSomething(void) {
-    int result = picobootOtpRead();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootGetInfo_should_DoSomething(void) {
-    int result = picobootGetInfo();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootPoke_should_DoSomething(void) {
-    int result = picobootPoke();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootPeek_should_DoSomething(void) {
-    int result = picobootPeek();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_picobootFlashId_should_DoSomething(void) {
-    int result = picobootFlashId();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
