--- test/test_port_windows.c
@@ -1,41 +1,0 @@
-#include "unity.h"
-#include "port_windows.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_portUartInit_should_DoSomething(void) {
-    int result = portUartInit();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_portUartWrite_should_DoSomething(void) {
-    int result = portUartWrite();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_portUartRead_should_DoSomething(void) {
-    int result = portUartRead();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_portFlushRx_should_DoSomething(void) {
-    portFlushRx();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_portMillis_should_DoSomething(void) {
-    int result = portMillis();
-    TEST_ASSERT_EQUAL(0, result);  // Replace 0 with expected return value
-}
-
-void test_portDelay_should_DoSomething(void) {
-    portDelay();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_portDebugPrintf_should_DoSomething(void) {
-    portDebugPrintf();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
