--- test/test_set_perms.c
@@ -1,11 +1,0 @@
-#include "unity.h"
-#include "set_perms.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_main_should_DoSomething(void) {
-    main();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
