--- test/test_test_uart.c
@@ -1,36 +1,0 @@
-#include "unity.h"
-#include "test_uart.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_setUp_should_DoSomething(void) {
-    setUp();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_tearDown_should_DoSomething(void) {
-    tearDown();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testUartInit_should_DoSomething(void) {
-    testUartInit();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testUartWrite_should_DoSomething(void) {
-    testUartWrite();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testUartRead_should_DoSomething(void) {
-    testUartRead();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testUartClearReceiveBuffer_should_DoSomething(void) {
-    testUartClearReceiveBuffer();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
