--- test/test_test_xbee.c
@@ -1,36 +1,0 @@
-#include "unity.h"
-#include "test_xbee.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_setUp_should_DoSomething(void) {
-    setUp();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_tearDown_should_DoSomething(void) {
-    tearDown();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testXbeeInit_should_DoSomething(void) {
-    testXbeeInit();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testXbeeSendFrame_should_DoSomething(void) {
-    testXbeeSendFrame();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testXbeeToString_should_DoSomething(void) {
-    testXbeeToString();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
-void test_testAtCommandToString_should_DoSomething(void) {
-    testAtCommandToString();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
