--- test/test_whereami.c
@@ -1,11 +1,0 @@
-#include "unity.h"
-#include "whereami.c"
-
-void setUp(void) {}
-void tearDown(void) {}
-
-void test_if_should_DoSomething(void) {
-    if();
-    // Add appropriate assertions to check the state changes or side effects
-}
-
