+++ test/test_xbee_lr.c
@@ -105,26 +105,101 @@ void test_XBeeLRGetDevEUI_should_return_string_on_valid_response(void) {
     TEST_ASSERT_EQUAL_STRING("0001020304050607", buffer);
 }
 
+
+// Test XBeeLRSetAppEUI with valid string
+void test_XBeeLRSetAppEUI_should_return_true_for_valid_input(void) {
+    const char* appEUI = "A1B2C3D4E5F60708";
+    uint8_t response[17] = {0};
+    uint8_t respLen = 0;
+    apiSendAtCommandAndGetResponse_ExpectAndReturn(&mockXbee, AT_AE, NULL, 8, response, &respLen, 5000, 17, API_SEND_SUCCESS);
+    TEST_ASSERT_TRUE(XBeeLRSetAppEUI(&mockXbee, appEUI));
+}
+
+// Test XBeeLRSetAppEUI with invalid string
+void test_XBeeLRSetAppEUI_should_return_false_for_invalid_input(void) {
+    const char* appEUI = "BADLENGTH";
+    TEST_ASSERT_FALSE(XBeeLRSetAppEUI(&mockXbee, appEUI));
+}
+
+// Test XBeeLRSetClass
+void test_XBeeLRSetClass_should_send_AT_LC_command(void) {
+    uint8_t response[4];
+    uint8_t responseLength = 0;
+    char classVal = 'A';
+    apiSendAtCommandAndGetResponse_ExpectAndReturn(&mockXbee, AT_LC, (const uint8_t*)&classVal, 1, response, &responseLength, 5000, sizeof(response), API_SEND_SUCCESS);
+    TEST_ASSERT_TRUE(XBeeLRSetClass(&mockXbee, classVal));
+}
+
+// Test XBeeLRSendData success path
+void test_XBeeLRSendData_should_send_and_wait_for_tx_status(void) {
+    XBeeLRPacket_t packet = {
+        .payload = (uint8_t*)"hi",
+        .payloadSize = 2,
+        .port = 1,
+        .ack = 0
+    };
+    mockXbee.frameIdCntr = 1;
+    mockXbee.txStatusReceived = true;
+    mockXbee.deliveryStatus = 0x00;
+    
+    // Expect apiSendFrame to succeed
+    apiSendFrame_ExpectAndReturn(&mockXbee, XBEE_API_TYPE_LR_TX_REQUEST, NULL, 5, API_SEND_SUCCESS);
+    // Allow XBeeLRProcess to return immediately
+    TEST_ASSERT_EQUAL_UINT8(0x00, XBeeLRSendData(&mockXbee, &packet));
+}
+
+// Test XBeeLRHandleTransmitStatus
+void test_XBeeLRHandleTransmitStatus_should_parse_and_set_flags(void) {
+    XBeeLRPacket_t packet = {0};
+    xbee_api_frame_t frame = {
+        .type = XBEE_API_TYPE_TX_STATUS,
+        .length = 3,
+        .data = {0x00, 0x01, 0x00}  // Frame ID, Status
+    };
+    mockXbee.ctable = NULL;
+
+    XBeeLRHandleTransmitStatus(&mockXbee, &frame);
+    TEST_ASSERT_TRUE(mockXbee.txStatusReceived);
+    TEST_ASSERT_EQUAL_UINT8(0x00, mockXbee.deliveryStatus);
+}
+
+// Test XBeeLRHandleRxPacket
+void test_XBeeLRHandleRxPacket_should_invoke_receive_callback(void) {
+    xbee_api_frame_t frame = {
+        .type = XBEE_API_TYPE_LR_RX_PACKET,
+        .length = 5,
+        .data = {0x81, 0x01, 0x02, 0x03, 0x04}
+    };
+    bool callbackInvoked = false;
+    void onReceive(XBee* self, void* data) { callbackInvoked = true; }
+
+    XBeeCTable ctable = {.OnReceiveCallback = onReceive};
+    mockXbee.ctable = &ctable;
+
+    XBeeLRHandleRxPacket(&mockXbee, &frame);
+    TEST_ASSERT_TRUE(callbackInvoked);
+}
+
 // Add more unit tests for:
 // - XBeeLRSetAppEUI
 // - XBeeLRSetClass
 // - XBeeLRSendData (with mocks)
 // - XBeeLRHandleTransmitStatus and XBeeLRHandleRxPacket (mock frame parsing)
 
-int main(void) {
-    UNITY_BEGIN();
+// int main(void) {
+//     UNITY_BEGIN();
 
-    RUN_TEST(test_XBeeLRInit_should_return_true_on_uart_success);
-    RUN_TEST(test_XBeeLRConnected_should_return_true_when_response_is_1);
-    RUN_TEST(test_XBeeLRConnected_should_return_false_on_error);
-    RUN_TEST(test_XBeeLRSetAppKey_should_fail_on_invalid_length);
-    RUN_TEST(test_XBeeLRSetAppKey_should_fail_on_conversion_error);
-    RUN_TEST(test_XBeeLRSetAppKey_should_pass_on_valid_input);
-    RUN_TEST(test_XBeeLRSetNwkKey_should_fail_on_bad_hex);
-    RUN_TEST(test_XBeeLRSetJoinRX1Delay_should_succeed_with_valid_value);
-    RUN_TEST(test_XBeeLRSetRX2Frequency_should_succeed);
-    RUN_TEST(test_XBeeLRSetChannelsMask_should_fail_on_odd_length);
-    RUN_TEST(test_XBeeLRGetDevEUI_should_return_string_on_valid_response);
+//     RUN_TEST(test_XBeeLRInit_should_return_true_on_uart_success);
+//     RUN_TEST(test_XBeeLRConnected_should_return_true_when_response_is_1);
+//     RUN_TEST(test_XBeeLRConnected_should_return_false_on_error);
+//     RUN_TEST(test_XBeeLRSetAppKey_should_fail_on_invalid_length);
+//     RUN_TEST(test_XBeeLRSetAppKey_should_fail_on_conversion_error);
+//     RUN_TEST(test_XBeeLRSetAppKey_should_pass_on_valid_input);
+//     RUN_TEST(test_XBeeLRSetNwkKey_should_fail_on_bad_hex);
+//     RUN_TEST(test_XBeeLRSetJoinRX1Delay_should_succeed_with_valid_value);
+//     RUN_TEST(test_XBeeLRSetRX2Frequency_should_succeed);
+//     RUN_TEST(test_XBeeLRSetChannelsMask_should_fail_on_odd_length);
+//     RUN_TEST(test_XBeeLRGetDevEUI_should_return_string_on_valid_response);
 
-    return UNITY_END();
-}
+//     return UNITY_END();
+// }
